Small fixes to client cod; clarify client IP info
At least in Python 3.10 I had a couple of errors: (1) missing NetGear import (2) can't return a value in an async. In addition, although it is clear in your several comments, I still failed to understand that the server uses the client's IP, so I reinforced this in the first info comment.